+++ python/cudf/cudf/core/buffer/__init__.py
@@ -1,10 +1,10 @@
 # Copyright (c) 2022, NVIDIA CORPORATION.
 
 from cudf.core.buffer.buffer import Buffer, cuda_array_interface_wrapper
+from cudf.core.buffer.cow_buffer import CopyOnWriteBuffer
 from cudf.core.buffer.spillable_buffer import SpillableBuffer, SpillLock
 from cudf.core.buffer.utils import (
     acquire_spill_lock,
     as_buffer,
     get_spill_lock,
 )
-from cudf.core.buffer.weakrefable_buffer import CopyOnWriteBuffer
